xclbinutil: enable critical compiler warnings (#10007)
* xclbinutil: enable critical compiler warnings


* xclbinutil: address clang-tidy diagnostics


* xclbinutil: address review feedback: apply patch pdi-transform-overflow-error-checking.patch


---------

+++ src/runtime_src/tools/xclbinutil/aie-pdi-transform/libinclude/cdo_cmd.h
@@ -106,7 +106,7 @@ typedef struct {
 typedef struct {
 	uint8_t bLeft;
 	uint32_t cmd_id;
-	uint16_t cmdNum;
+	uint32_t cmdNum;
 	uint32_t cmdHeaderLeft[MAX_HEADER_LEFT];
 	uint32_t cmdHeaderLeftLen;
 } XCdoCmdLeft;
